Implement sorting in ToDo and ToDoDatabase
Not sure about quality of solution- had to add getters in ToDo.java, which we're not sure was a good idea as it affects the simplicity of the object definition class.

We don't have tests written at the moment, but it appears to be working correctly. May add tests later.